// src/shared/api/generated/models/endUserCasePolicyPreviewResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { EndUserCasePolicyPreviewResponseDtoCompiledPolicy } from "./endUserCasePolicyPreviewResponseDtoCompiledPolicy";
import type { EndUserCasePolicyPreviewExampleResponseDto } from "./endUserCasePolicyPreviewExampleResponseDto";

export interface EndUserCasePolicyPreviewResponseDto {
  compiledPolicy: EndUserCasePolicyPreviewResponseDtoCompiledPolicy;
  compiledPolicyHash: string;
  compilerVersion: string;
  examples: EndUserCasePolicyPreviewExampleResponseDto[];
}